Add "Sample" TEE for KBS-AS tests.
In some scenarios, KBS links do not want to be based on real TEE authentication,
but only want to test the normal functions of KBS https server.
This is because the test environment may not be a real TEE environment.
This "Sample" TEE type is for the basic test of KBS.

Signed-off-by: Jiale Zhang <zhangjiale@linux.alibaba.com>