Add new features to use this crate without the standard library
It can be useful to use this crate in bare-metal code where `std`
is not available (e.g. SVSM).

So let's provide 2 new features (`std` and `alloc`) to support this
use case. `std` is enabled by default.

As long as there is a memory allocator, we can still use this crate
without the rest of the standard library by enabling the `alloc`
feature:

  `kbs-types = { version = "...", default-features = false,
                 features = ["alloc"] }`

The patch is based on what serde does to handle the same use case.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>