Apply RFC7516 Standard to Response
This patch is a fix up to previous commits to update the Response struct
to follow RFC7516.

The work included in the commit
1. Make both serialization and deserialization of `ProtectedHeader`
use base64-url-nopadding encoding. This means that the whole
ProtectedHeader will be serialized into a base64 string, rather than a
struct.
2. Make `encrypted_key`, `aad`, `iv`, `ciphertext`, `tag` to
base64-url-nopadding encoded string when serializing.
3. Automatical deseralization logic when parsing Response from a
flattened JSON JWE Serialization.

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>