Attestation: add runtime and initdata field
This is to allow kbs-protocol level to have two high level abstraction,
s.t.

things needed to be binded to evidence at runtime: runtime data
things already binded to evidence at initialization of TEE: initdata

The legal nonce and tee public key are both runtime data, thus we still
leave the definition of tee public key but we do not require it inside
the runtime data field, it is left to implementers to decide.

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>